Fix resizing of hidden elements
`_watchSizes` returns the current size of an element. It returns `undefined` if the element in question is hidden, which makes sense. However, the result of `_watchSizes` is always assigned to `oldSizes`. The intent of variable `oldSizes` is to keep track of the previous size of an element, so the code can check if there's a difference in size later on. `oldSizes` should never be `undefined`. This change makes sure of that.